- Problem 12
- Problem 20
- Problem 22
- Problem 23
- Problem 24
- Problem 26
- Problem 27
- Problem 28
- Problem 29
- Problem 30
- Problem 31
- Problem 32

+++ src/_3_problems_from_20_to_55/_3_1_problems_from_20_to_32/_3_1_12_problem_31/Problem31.cpp
@@ -177,7 +177,7 @@ void nextYear(
 void nextDecade(
     Date& date
 ) {
-    for (int year = 0; year < 10; ++year)
+    for (short year = 0; year < 10; ++year)
         nextYear(
             date
         );
@@ -186,7 +186,7 @@ void nextDecade(
 void nextCentury(
     Date& date
 ) {
-    for (int decades = 0; decades < 10; ++decades)
+    for (short decades = 0; decades < 10; ++decades)
         nextDecade(
             date
         );
